feat: update for 6.35.1
Significant changes have happened since this was last cross-compiled,
but with a bit more elbow-grease it seems to be working again.

Most important things: enabling XSPEC models will only compile heacore
and Xspec. So really no need to patch all the other libraries in the
future.

Default Apple build is now for aarch64, so need to enable a flag when
compiling for Intel-based macs.